Sixth major checkpoint to SwiftUI-fy the package
