+++ Sources/Pixel/Core/Colors/ColorSchemes/Extensions/PixelColorScheme+Empty.swift
@@ -1,0 +1,35 @@
+//
+//  PixelColorScheme+Empty.swift
+//  Pixel
+//
+//  Created by Khaled Chehabeddine on 31/10/2024.
+//  Copyright © 2024 Pixel. All rights reserved.
+//
+
+import Foundation
+
+public extension PixelColorScheme {
+
+    static let empty: Self = .init(
+        primary: .empty,
+        onPrimary: .empty,
+        secondary: .empty,
+        onSecondary: .empty,
+        accent: .empty,
+        onAccent: .empty,
+        neutral: .empty,
+        onNeutral: .empty,
+        background100: .empty,
+        background200: .empty,
+        background300: .empty,
+        onBackground: .empty,
+        info: .empty,
+        onInfo: .empty,
+        error: .empty,
+        onError: .empty,
+        warning: .empty,
+        onWarning: .empty,
+        success: .empty,
+        onSuccess: .empty
+    )
+}
