--- Sources/Pixel/Core/Colors/Extensions/Color/Color+Pixel.swift
@@ -1,15 +1,0 @@
-//
-//  Color+Pixel.swift
-//  Pixel
-//
-//  Created by Khaled Chehabeddine on 18/08/2024.
-//  Copyright © 2024 Pixel. All rights reserved.
-//
-
-import SwiftUI
-
-public extension Color {
-
-    enum Pixel {
-    }
-}
