+++ Sources/Pixel/Core/Colors/Extensions/PixelColor+DarkColorScheme.swift
@@ -1,0 +1,14 @@
+//
+//  PixelColor+DarkColorScheme.swift
+//  Pixel
+//
+//  Created by Khaled Chehabeddine on 31/10/2024.
+//  Copyright © 2024 Pixel. All rights reserved.
+//
+
+import Foundation
+
+public extension PixelColor {
+
+    static let darkColorScheme: PixelColorScheme = .dark
+}
