+++ Sources/Pixel/Core/Colors/Extensions/PixelColor+LightColorScheme.swift
@@ -1,0 +1,14 @@
+//
+//  PixelColor+LightColorScheme.swift
+//  Pixel
+//
+//  Created by Khaled Chehabeddine on 31/10/2024.
+//  Copyright © 2024 Pixel. All rights reserved.
+//
+
+import Foundation
+
+public extension PixelColor {
+
+    static let lightColorScheme: PixelColorScheme = .light
+}
