--- Sources/Pixel/Core/Colors/Extensions/UIColor/UIColor+Pixel.swift
@@ -1,15 +1,0 @@
-//
-//  UIColor+Pixel.swift
-//  Pixel
-//
-//  Created by Khaled Chehabeddine on 18/08/2024.
-//  Copyright © 2024 Pixel. All rights reserved.
-//
-
-import UIKit
-
-public extension UIColor {
-
-    enum Pixel {
-    }
-}
