--- Sources/Pixel/Core/Colors/Extensions/UIColor/UIColor.Pixel+Dark.swift
@@ -1,41 +1,0 @@
-//
-//  UIColor.Pixel+Dark.swift
-//  Pixel
-//
-//  Created by Khaled Chehabeddine on 18/08/2024.
-//  Copyright © 2024 Pixel. All rights reserved.
-//
-
-import UIKit
-
-public extension UIColor.Pixel {
-
-    enum Dark {
-
-        public static let primary: UIColor = PixelColorScheme.dark.primary.uiColor
-        public static let onPrimary: UIColor = PixelColorScheme.dark.onPrimary.uiColor
-
-        public static let secondary: UIColor = PixelColorScheme.dark.secondary.uiColor
-        public static let onSecondary: UIColor = PixelColorScheme.dark.onSecondary.uiColor
-
-        public static let accent: UIColor = PixelColorScheme.dark.accent.uiColor
-        public static let onAccent: UIColor = PixelColorScheme.dark.onAccent.uiColor
-
-        public static let neutral: UIColor = PixelColorScheme.dark.neutral.uiColor
-        public static let onNeutral: UIColor = PixelColorScheme.dark.onNeutral.uiColor
-
-        public static let background100: UIColor = PixelColorScheme.dark.background100.uiColor
-        public static let background200: UIColor = PixelColorScheme.dark.background200.uiColor
-        public static let background300: UIColor = PixelColorScheme.dark.background300.uiColor
-        public static let onBackground: UIColor = PixelColorScheme.dark.onBackground.uiColor
-
-        public static let info: UIColor = PixelColorScheme.dark.info.uiColor
-        public static let onInfo: UIColor = PixelColorScheme.dark.onInfo.uiColor
-        public static let error: UIColor = PixelColorScheme.dark.error.uiColor
-        public static let onError: UIColor = PixelColorScheme.dark.onError.uiColor
-        public static let warning: UIColor = PixelColorScheme.dark.warning.uiColor
-        public static let onWarning: UIColor = PixelColorScheme.dark.onWarning.uiColor
-        public static let success: UIColor = PixelColorScheme.dark.success.uiColor
-        public static let onSuccess: UIColor = PixelColorScheme.dark.onSuccess.uiColor
-    }
-}
