--- Sources/Pixel/Core/Colors/Extensions/UIColor/UIColor.Pixel+Light.swift
@@ -1,41 +1,0 @@
-//
-//  UIColor.Pixel+Light.swift
-//  Pixel
-//
-//  Created by Khaled Chehabeddine on 19/08/2024.
-//  Copyright © 2024 Pixel. All rights reserved.
-//
-
-import UIKit
-
-public extension UIColor.Pixel {
-
-    enum Light {
-
-        public static let primary: UIColor = PixelColorScheme.light.primary.uiColor
-        public static let onPrimary: UIColor = PixelColorScheme.light.onPrimary.uiColor
-
-        public static let secondary: UIColor = PixelColorScheme.light.secondary.uiColor
-        public static let onSecondary: UIColor = PixelColorScheme.light.onSecondary.uiColor
-
-        public static let accent: UIColor = PixelColorScheme.light.accent.uiColor
-        public static let onAccent: UIColor = PixelColorScheme.light.onAccent.uiColor
-
-        public static let neutral: UIColor = PixelColorScheme.light.neutral.uiColor
-        public static let onNeutral: UIColor = PixelColorScheme.light.onNeutral.uiColor
-
-        public static let background100: UIColor = PixelColorScheme.light.background100.uiColor
-        public static let background200: UIColor = PixelColorScheme.light.background200.uiColor
-        public static let background300: UIColor = PixelColorScheme.light.background300.uiColor
-        public static let onBackground: UIColor = PixelColorScheme.light.onBackground.uiColor
-
-        public static let info: UIColor = PixelColorScheme.light.info.uiColor
-        public static let onInfo: UIColor = PixelColorScheme.light.onInfo.uiColor
-        public static let error: UIColor = PixelColorScheme.light.error.uiColor
-        public static let onError: UIColor = PixelColorScheme.light.onError.uiColor
-        public static let warning: UIColor = PixelColorScheme.light.warning.uiColor
-        public static let onWarning: UIColor = PixelColorScheme.light.onWarning.uiColor
-        public static let success: UIColor = PixelColorScheme.light.success.uiColor
-        public static let onSuccess: UIColor = PixelColorScheme.light.onSuccess.uiColor
-    }
-}
