+++ Sources/Pixel/Core/Fonts/PixelFont/Extensions/AnyPixelFont+SFPro.swift
@@ -1,0 +1,14 @@
+//
+//  AnyPixelFont+SFPro.swift
+//  Pixel
+//
+//  Created by Khaled Chehabeddine on 31/10/2024.
+//  Copyright © 2024 Pixel. All rights reserved.
+//
+
+import Foundation
+
+public extension AnyPixelFont {
+
+    static let sfProTypography: PixelTypography = .sfPro
+}
