+++ Sources/Pixel/Core/Fonts/Typography/Extensions/PixelTypography+Empty.swift
@@ -1,0 +1,45 @@
+//
+//  PixelTypography+Empty.swift
+//  Pixel
+//
+//  Created by Khaled Chehabeddine on 31/10/2024.
+//  Copyright © 2024 Pixel. All rights reserved.
+//
+
+import Foundation
+
+public extension PixelTypography {
+
+    static let empty: Self = .init(
+        superDino1: .empty,
+        superDino2: .empty,
+        superDino3: .empty,
+        superDino4: .empty,
+        capsSuperDino: .empty,
+        dino1: .empty,
+        dino2: .empty,
+        dino3: .empty,
+        dino4: .empty,
+        capsDino: .empty,
+        big1: .empty,
+        big2: .empty,
+        big3: .empty,
+        big4: .empty,
+        capsBig: .empty,
+        medium1: .empty,
+        medium2: .empty,
+        medium3: .empty,
+        medium4: .empty,
+        capsMedium: .empty,
+        small1: .empty,
+        small2: .empty,
+        small3: .empty,
+        small4: .empty,
+        capsSmall: .empty,
+        superSmall1: .empty,
+        superSmall2: .empty,
+        superSmall3: .empty,
+        superSmall4: .empty,
+        capsSuperSmall: .empty
+    )
+}
