+++ Sources/Pixel/Core/Fonts/Typography/Extensions/PixelTypography+SFPro.swift
@@ -1,5 +1,5 @@
 //
-//  PixelFont+PixelTypography.swift
+//  PixelTypography+SFPro.swift
 //  Pixel
 //
 //  Created by Khaled Chehabeddine on 21/08/2024.
@@ -59,68 +59,35 @@ private extension SFProPixelFont {
 public extension PixelTypography {
 
     static let sfPro: Self = .init(
-        superDino1: .init(SFProPixelFont.superDino1),
-        superDino2: .init(SFProPixelFont.superDino2),
-        superDino3: .init(SFProPixelFont.superDino3),
-        superDino4: .init(SFProPixelFont.superDino4),
-        capsSuperDino: .init(SFProPixelFont.capsSuperDino),
-        dino1: .init(SFProPixelFont.dino1),
-        dino2: .init(SFProPixelFont.dino2),
-        dino3: .init(SFProPixelFont.dino3),
-        dino4: .init(SFProPixelFont.superDino1),
-        capsDino: .init(SFProPixelFont.capsDino),
-        big1: .init(SFProPixelFont.big1),
-        big2: .init(SFProPixelFont.big2),
-        big3: .init(SFProPixelFont.big3),
-        big4: .init(SFProPixelFont.big4),
-        capsBig: .init(SFProPixelFont.capsBig),
-        medium1: .init(SFProPixelFont.medium1),
-        medium2: .init(SFProPixelFont.medium2),
-        medium3: .init(SFProPixelFont.medium3),
-        medium4: .init(SFProPixelFont.medium4),
-        capsMedium: .init(SFProPixelFont.capsMedium),
-        small1: .init(SFProPixelFont.small1),
-        small2: .init(SFProPixelFont.small2),
-        small3: .init(SFProPixelFont.small3),
-        small4: .init(SFProPixelFont.small4),
-        capsSmall: .init(SFProPixelFont.capsSmall),
-        superSmall1: .init(SFProPixelFont.superSmall1),
-        superSmall2: .init(SFProPixelFont.superSmall2),
-        superSmall3: .init(SFProPixelFont.superSmall3),
-        superSmall4: .init(SFProPixelFont.superSmall4),
-        capsSuperSmall: .init(SFProPixelFont.capsSuperSmall)
-    )
-
-    static let empty: Self = .init(
-        superDino1: .empty,
-        superDino2: .empty,
-        superDino3: .empty,
-        superDino4: .empty,
-        capsSuperDino: .empty,
-        dino1: .empty,
-        dino2: .empty,
-        dino3: .empty,
-        dino4: .empty,
-        capsDino: .empty,
-        big1: .empty,
-        big2: .empty,
-        big3: .empty,
-        big4: .empty,
-        capsBig: .empty,
-        medium1: .empty,
-        medium2: .empty,
-        medium3: .empty,
-        medium4: .empty,
-        capsMedium: .empty,
-        small1: .empty,
-        small2: .empty,
-        small3: .empty,
-        small4: .empty,
-        capsSmall: .empty,
-        superSmall1: .empty,
-        superSmall2: .empty,
-        superSmall3: .empty,
-        superSmall4: .empty,
-        capsSuperSmall: .empty
+        superDino1: SFProPixelFont.superDino1.eraseToAnyPixelFont,
+        superDino2: SFProPixelFont.superDino2.eraseToAnyPixelFont,
+        superDino3: SFProPixelFont.superDino3.eraseToAnyPixelFont,
+        superDino4: SFProPixelFont.superDino4.eraseToAnyPixelFont,
+        capsSuperDino: SFProPixelFont.capsSuperDino.eraseToAnyPixelFont,
+        dino1: SFProPixelFont.dino1.eraseToAnyPixelFont,
+        dino2: SFProPixelFont.dino2.eraseToAnyPixelFont,
+        dino3: SFProPixelFont.dino3.eraseToAnyPixelFont,
+        dino4: SFProPixelFont.superDino1.eraseToAnyPixelFont,
+        capsDino: SFProPixelFont.capsDino.eraseToAnyPixelFont,
+        big1: SFProPixelFont.big1.eraseToAnyPixelFont,
+        big2: SFProPixelFont.big2.eraseToAnyPixelFont,
+        big3: SFProPixelFont.big3.eraseToAnyPixelFont,
+        big4: SFProPixelFont.big4.eraseToAnyPixelFont,
+        capsBig: SFProPixelFont.capsBig.eraseToAnyPixelFont,
+        medium1: SFProPixelFont.medium1.eraseToAnyPixelFont,
+        medium2: SFProPixelFont.medium2.eraseToAnyPixelFont,
+        medium3: SFProPixelFont.medium3.eraseToAnyPixelFont,
+        medium4: SFProPixelFont.medium4.eraseToAnyPixelFont,
+        capsMedium: SFProPixelFont.capsMedium.eraseToAnyPixelFont,
+        small1: SFProPixelFont.small1.eraseToAnyPixelFont,
+        small2: SFProPixelFont.small2.eraseToAnyPixelFont,
+        small3: SFProPixelFont.small3.eraseToAnyPixelFont,
+        small4: SFProPixelFont.small4.eraseToAnyPixelFont,
+        capsSmall: SFProPixelFont.capsSmall.eraseToAnyPixelFont,
+        superSmall1: SFProPixelFont.superSmall1.eraseToAnyPixelFont,
+        superSmall2: SFProPixelFont.superSmall2.eraseToAnyPixelFont,
+        superSmall3: SFProPixelFont.superSmall3.eraseToAnyPixelFont,
+        superSmall4: SFProPixelFont.superSmall4.eraseToAnyPixelFont,
+        capsSuperSmall: SFProPixelFont.capsSuperSmall.eraseToAnyPixelFont
     )
 }
