--- Sources/Pixel/Pixel.swift
@@ -1,12 +1,0 @@
-//
-//  Pixel.swift
-//  Pixel
-//
-//  Created by Khaled Chehabeddine on 18/08/2024.
-//  Copyright © 2024 Pixel. All rights reserved.
-//
-
-import Foundation
-
-public enum Pixel {
-}
